chore: textvariant 이관, reset css 수정
